Домашка 2.3 доработка

+++ src/main/java/pro/sky/java/course2/dz2_3_1/service/CalculatorServiceIml.java
@@ -5,13 +5,13 @@ import org.springframework.stereotype.Service;
 @Service
 public class CalculatorServiceIml{
 
-    public int plus(int num1, int num2) {
+    public double plus(double num1, double num2) {
         return num1 + num2;
     }
-    public int minus(int num1, int num2) {
+    public double minus(double num1, double num2) {
         return num1 - num2;
     }
-    public int multiply(int num1, int num2) {
+    public double multiply(double num1, double num2) {
         return num1 * num2;
     }
     public double divide(double num1, double num2) {
